Move low level library stuff into RAM.
bank switching for every little string routine or conio call was getting tiresome.
Now, the linker binds that stuff into top 8k of expansion ram,
and the makefile appends that to the end of bank_1. A routine in bank_1 memcpys it
up into ram. This is the first routine called by main.