Move ea5 launch instruction to rom header
Fix so they are in stable address no matter what kind of
bank switching may have been erroniously performed by target app.

Fix so WS at time of launch is GPLWS. Found a program that assumes
that and stores the GPLWS R11 for return on exit, instead of the
current workspace. That is fair enough.

EDIT80 loads and exits fine now.