all the bits from libti99 in ROM
Nothing will work yet, but it's setup to be optimized as necessary,
slotted into appropriate banks if only used in a few places...
most of libti99 is the conio stuff and screen setup, that can all
co-locate in bank8 with the terminal screen code.

Next are string utilities... they'll go in bank1, and need to be
modified to be callable from other banks.

Probably should just inline the low level VDP routines.