configure sams memory if available
Also add message and halt if no expansion memory is present.

SAMS is optional.
When sams is used, we need to set the mapper up before we start using
any of the memory. So, stack is in scratchpad temporarily. Then sams
is detected, and 8 pages mapped into memory expansion space.
The the initial state is loaded from ROM, and block data is cleared.

Sams mapper is always on once detected.