protect against getting too close to the stack
users of alloc should check for 0 and handle failure to allocate.
currently the only other use is in building the DSR entry list, after
which there is typically still 4k of stack/heap available.

Added bk_free to restore heap to a previous value.

alloc and free really only work if the alloc is perminent, or
you are freeing the last things allocated.