fix delete to work on IDE
Needed to close the CATALOG before mutating it with delete operations.
So now we build out a list of matching names to delete into the heap.
then close the catalog, and iterate over the names and delete them.

probably a good pattern for the other wildcarding operations. However
those that simply read do not seem to get errors.