fix adding labels as file parses
when I changed to the global goto_line_ref while fixing
nested scripts I failed to back out a partial implementation.

consequently nobody was passing the current line number to
handleCommand, and so garbage was used.

This exposes, that if there is a problem in the banked function
signature disagreeing with the natural function signature, the
casting will hide it from the compiler... ick..