all target filename for copy command
if the target is an existing directory, the file will be copied their
with same name.

if the target is a non-existant thing, then the file will be copied to
the parent dir of the target and with the new name in the target.

if wildcarding in the source, then the target can only be a directory.

fix issue #26