script from menu command can cause exit
FCM=QUIT in a script run by FCMENU will
instruct FCMENU to quit upon return after
the script is complete.

FCM will be unset once consumed.

This allows creating menu entries that setup
a shell environment for a task, and then drop
you into that Force Command interactive mode.

A little bit of a hack, but it gets the job
done.