provide libgcc double/float support
The functions the compiler may call for float and double support
of add, subtract, multiply and divide are now in bank0 of the
cartridge.

Branch table entries are provided in the fc.ld linkerscript
so executables do not link directly to the cartridge.

To get these routines, I copyied the double.c from tms9900 gcc's
libgcc. I've optimized it to not create data segment trash for
pointers, and to use workspace >8300, consistent with Force Command.